added tags

+++ cogs/tags.py
@@ -1,5 +1,6 @@
 import discord
 import os
+import os.path
 import json
 from discord.ext import commands
 from discord import app_commands
@@ -22,16 +23,21 @@ class Tags(commands.Cog):
         if not os.path.exists("tags"):
             os.makedirs("tags")
 
-        embed_dict = {
-            "name": name,
-            "creator": interaction.user.name,
-            "content": content
-        }
+        path = f"tags/{name}.json"
+        fileExists = os.path.exists(path)
+        if fileExists:
+            embed_dict = {
+                "name": name,
+                "creator": interaction.user.name,
+                "content": content
+            }
 
-        with open(f'tags/{name}.json', 'w') as f:
-            json.dump(embed_dict, f, indent=4)
-        print(f"{interaction.user} created a tag")
-        await interaction.response.send_message(f"Tag {name} created!")
+            with open(f'tags/{name}.json', 'w') as f:
+                json.dump(embed_dict, f, indent=4)
+            print(f"{interaction.user} created a tag")
+            await interaction.response.send_message(f"Tag {name} created!")
+        else:
+            await interaction.response.send_message(f"Tag {name} already exists!", ephemeral=True)
     
     @app_commands.command(name='tag', description='Read a tag')
     @app_commands.guilds(discord.Object(id=server_id))
@@ -47,6 +53,18 @@ class Tags(commands.Cog):
             await interaction.response.send_message(f"`tags/{tag}.json` || {content}\n-# tag created by {creator}")
         except FileNotFoundError:
             await interaction.response.send_message(f"***[Errno 1]***: tags/{tagname}.json does not exist")
+            
+    @app_commands.command(name='delete-tag', description='Delete a tag')
+    @app_commands.guilds(discord.Object(id=server_id))
+    @discord.app_commands.checks.has_permissions(manage_messages=True)
+    async def deleteTag(self, interaction: discord.Interaction, tag: str, reason: str):
+        try:
+            os.remove(f"tags/{tag}.json")
+            embed = discord.Embed(title=f"Removed tag {tag}", description=f"Reason:\n{reason}", colour=0xff7800)
+            embed.set_author(name=interaction.user.name, icon_url=interaction.user.avatar)
+            await interaction.response.send_message(embed=embed)
+        except FileNotFoundError:
+            await interaction.response.send_message(f"***[Errno 1]***: File not found")
 
 async def setup(bot: commands.Bot):
     await bot.add_cog(Tags(bot))
